zonelan_backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
mysqlclient>=2.0,<3.0
django-cors-headers>=3.10.0
djangorestframework-simplejwt>=5.0.0
python-dotenv>=0.19.0